👌 Move NS sync feature to ArgoCD pkg
