+++ upgrades/0.0.97.seperate-ns-from-app/pkg/argocd/testdata/argoapp-Should setup correct files and directories.golden
