+++ upgrades/0.0.97.seperate-ns-from-app/pkg/upgrade/api.go
@@ -5,6 +5,9 @@ import (
 	"fmt"
 	"path"
 
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.97.seperate-ns-from-app/pkg/argocd"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.97.seperate-ns-from-app/pkg/paths"
+
 	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.97.seperate-ns-from-app/pkg/lib/manifest/apis/okctl.io/v1alpha1"
 	"github.com/spf13/afero"
 
@@ -13,14 +16,14 @@ import (
 )
 
 func Start(_ context.Context, logger logging.Logger, fs *afero.Afero, flags cmdflags.Flags, cluster v1alpha1.Cluster) error {
-	absoluteRepositoryRootDir, err := getRepositoryRootDirectory()
+	absoluteRepositoryRootDir, err := paths.GetRepositoryRootDirectory()
 	if err != nil {
 		return fmt.Errorf("acquiring repository root dir: %w", err)
 	}
 
-	absoluteNamespacesDir := path.Join(absoluteRepositoryRootDir, namespacesDir(cluster))
+	absoluteNamespacesDir := path.Join(absoluteRepositoryRootDir, paths.RelativeNamespacesDir(cluster))
 
-	err = setupNamespacesSync(fs, cluster)
+	err = argocd.EnableNamespacesSync(fs, cluster)
 	if err != nil {
 		return fmt.Errorf("adding namespaces app manifest: %w", err)
 	}
